Fix offset off-by-one at end of input
The read() method fails to increment the index when reading the last
character of the input.
Other methods compensate for this by treating the index differently
when at end of text.
By correctly incrementing the index also when at end of text, the
conditional adjustments can be removed from endCapture() and
getLocation().