Change JSON parser to start column numbers at one
Most editors regard the first character position of a text as "1:1".
his seems more logical than "1:0" which is currently returned for the
first position by icecore-json.
Adapt this established pattern and let column numbers start at 1, as
line numbers do.

With this change, there is an easy rule:
  - "offset" is for programs, it starts at 0
  - "line" and "column" are for humans, they start at 1

This leads to a slight API change in ParseException.
However, since the effect can only be seen in the case of an error,
it seems acceptable.